Eliminated compilation warning.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@12054 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/aws/awstxtbx.cpp
@@ -340,10 +340,10 @@ void awsTextBox::OnDraw (csRect /*clip*/)
     if (mcc)
     {
 
-      // Check to see if we're getting wierd.
+      // Check to see if we're getting weird.
       // this was changed to avoid 
       // jittering in the start value
-      if (cursor > start + mcc) start = cursor - mcc;
+      if ((int)cursor > start + mcc) start = cursor - mcc;
       if (start < 0) start = 0;
 
       // Make the text the right length
